Temp solution to multi-background problem
