+++ packages/yoroi-extension/chrome/extension/background.js
@@ -1,13 +1,6 @@
 // @flow
 import debounce from 'lodash/debounce';
 
-import { schema } from 'lovefield';
-import type {
-  lf$Database,
-} from 'lovefield';
-import {
-  loadLovefieldDB,
-} from '../../app/api/ada/lib/storage/database/index';
 import {
   getWallets
 } from '../../app/api/common/index';
@@ -16,13 +9,10 @@ import {
 } from '../../app/api/ada/lib/storage/models/PublicDeriver/index';
 import {
   asGetAllUtxos,
-  asGetPublicKey,
 } from '../../app/api/ada/lib/storage/models/PublicDeriver/traits';
 import type {
   PendingSignData,
   PendingTransaction,
-  RpcUid,
-  AccountInfo,
 } from './ergo-connector/types';
 import {
   APIErrorCodes,
@@ -42,11 +32,6 @@ import {
   connectorGetUsedAddresses,
   connectorGetUnusedAddresses
 } from './ergo-connector/api';
-import { GenericApiError } from '../../app/api/common/errors';
-import { isErgo, isCardanoHaskell, } from '../../app/api/ada/lib/storage/database/prepackaged/networks';
-import { Bip44Wallet } from '../../app/api/ada/lib/storage/models/Bip44Wallet/wrapper';
-import { walletChecksum, legacyWalletChecksum } from '@emurgo/cip4-js';
-import type { WalletChecksum } from '@emurgo/cip4-js';
 import { updateTransactions } from '../../app/api/ergo/lib/storage/bridge/updateTransactions';
 import { environment } from '../../app/environment';
 import { IFetcher } from '../../app/api/ergo/lib/state-fetch/IFetcher';
@@ -55,6 +40,8 @@ import { BatchedFetcher } from '../../app/api/ergo/lib/state-fetch/batchedFetche
 import LocalStorageApi from '../../app/api/localStorage/index';
 import { RustModule } from '../../app/api/ada/lib/cardanoCrypto/rustLoader';
 import { Logger } from '../../app/utils/logging';
+import { loadDB } from './utils';
+import type { AccountIndex, ConnectedSite } from './utils';
 
 /*::
 declare var chrome;
@@ -66,101 +53,12 @@ const onYoroiIconClicked = () => {
 
 chrome.browserAction.onClicked.addListener(debounce(onYoroiIconClicked, 500, { leading: true }));
 
-type PendingSign = {|
-  // data needed to complete the request
-  request: PendingSignData,
-  // if an opened window has been created for this request to show the user
-  openedWindow: boolean,
-  // resolve function from signRequest's Promise
-  resolve: any
-|}
-
-// This is a temporary workaround to DB duplicate key constraint violations
-// that is happening when multiple DBs are loaded at the same time, or possibly
-// this one being loaded while Yoroi's main App is doing DB operations.
-let loadedDB: ?lf$Database = null;
-let dbPromise: ?Promise<lf$Database> = null;
-
-async function loadDB(): Promise<lf$Database> {
-  if (loadedDB == null) {
-    if (dbPromise == null) {
-      dbPromise = loadLovefieldDB(schema.DataStoreType.INDEXED_DB)
-        .then(db => {
-          loadedDB = db;
-          return Promise.resolve(loadedDB);
-        });
-    }
-    return dbPromise;
-  }
-  return Promise.resolve(loadedDB);
-}
-
-type AccountIndex = number;
-
-// AccountIndex = successfully connected - which account the user selected
-// null = refused by user
-type ConnectedStatus = ?AccountIndex | {|
-  // response (?AccountIndex) - null means the user refused, otherwise the account they selected
-  resolve: any,
-  // if a window has fetched this to show to the user yet
-  openedWindow: boolean,
-|};
-
-type ConnectedSite = {|
-  url: string,
-  status: ConnectedStatus,
-  pendingSigns: Map<RpcUid, PendingSign>
-|};
-
-async function getChecksum(
-  publicDeriver: ReturnType<typeof asGetPublicKey>,
-): Promise<void | WalletChecksum> {
-  if (publicDeriver == null) return undefined;
-
-  const hash = (await publicDeriver.getPublicKey()).Hash;
-
-  const isLegacyWallet =
-    isCardanoHaskell(publicDeriver.getParent().getNetworkInfo()) &&
-    publicDeriver.getParent() instanceof Bip44Wallet;
-  const checksum = isLegacyWallet
-    ? legacyWalletChecksum(hash)
-    : walletChecksum(hash);
-
-  return checksum;
-}
-
 // tab id key
 const connectedSites: Map<number, ConnectedSite> = new Map();
 
 let pendingTxs: PendingTransaction[] = [];
 
-export async function getWalletsInfo(): Promise<AccountInfo[]> {
-  try {
-    const db = await loadDB();
-    const wallets = await getWallets({ db });
-    // information about each wallet to show to the user
-    const accounts = [];
-    for (const wallet of wallets) {
-      const conceptualWallet = wallet.getParent();
-      const withPubKey = asGetPublicKey(wallet);
-
-      const conceptualInfo = await conceptualWallet.getFullConceptualWalletInfo();
-      if (isErgo(conceptualWallet.getNetworkInfo())) {
-        const balance = await connectorGetBalance(wallet, pendingTxs, 'ERG');
-        accounts.push({
-          name: conceptualInfo.Name,
-          balance: balance.toString(),
-          checksum: await getChecksum(withPubKey)
-        });
-      }
-    }
-    return accounts;
-  } catch (error) {
-    throw new GenericApiError();
-  }
-}
-
-export async function getStateFetcher(): Promise<IFetcher> {
+async function getStateFetcher(): Promise<IFetcher> {
   // I don't think it's worth it to cache this? We only need it for syncs and sending tx
   const localStorgeApi = new LocalStorageApi();
   const locale = await localStorgeApi.getUserLocale() ?? 'en-US';
@@ -221,13 +119,6 @@ async function syncWallet(wallet: PublicDeriver<>): Promise<void> {
   } catch (e) {
     Logger.error(`Syncing failed: ${e}`);
   }
-}
-export function getActiveSites(): Array<string> {
-  const activeSites = []
-  for (const value of connectedSites.values()){
-    activeSites.push(value);
-  }
-  return activeSites.map(item => item.url);
 }
 
 async function getSelectedWallet(tabId: number): Promise<PublicDeriver<>> {
